Bug 1626656 - Part 4: Add explicit casts to printf arguments to match either %d or %u specifier. r=jandem
Differential Revision: https://phabricator.services.mozilla.com/D69221

+++ js/src/gc/Statistics.cpp
@@ -344,7 +344,7 @@ UniqueChars Statistics::formatCompactSummaryMessage() const {
                  zoneStats.sweptZoneCount, zoneStats.collectedCompartmentCount,
                  zoneStats.compartmentCount, zoneStats.sweptCompartmentCount,
                  double(preTotalHeapBytes) / bytesPerMiB,
-                 counts[COUNT_NEW_CHUNK] - counts[COUNT_DESTROY_CHUNK],
+                 int32_t(counts[COUNT_NEW_CHUNK] - counts[COUNT_DESTROY_CHUNK]),
                  counts[COUNT_NEW_CHUNK] + counts[COUNT_DESTROY_CHUNK]);
   if (!fragments.append(DuplicateString(buffer))) {
     return UniqueChars(nullptr);
